// src/app/jobs/types/job-data.ts
export interface JobData {
  id: string;
  title: string;
  zip_code: string;
  city: string;
  thumbnail: string;
  attachments: string[];
  counter: {
    messages_total: number,
    messages_unread: number
  };
  is_awarded: boolean;
  categories: {id: string}[];
  state: string;
  description: string;
  end_date: string;
  created_at: string;
}